test: remove createTestEnv from Offer, merge Offer2 test into Ofer
